# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for BrandishBot game engine - inventory, crafting, economy, and stats management",
        "title": "BrandishBot API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/osse101/BrandishBot_Go"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/user/item/disassemble": {
            "post": {
                "description": "Disassemble an item into materials\nDisassemble an item into materials",
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "crafting",
                    "crafting"
                ],
                "summary": "Disassemble item",
                "parameters": [
                    {
                        "description": "Disassemble details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DisassembleItemRequest"
                        }
                    },
                    {
                        "description": "Disassemble details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DisassembleItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DisassembleItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Feature locked",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.DisassembleItemRequest": {
            "type": "object",
            "required": [
                "item",
                "username"
            ],
            "properties": {
                "item": {
                    "type": "string",
                    "maxLength": 100
                },
                "platform": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "handler.DisassembleItemResponse": {
            "type": "object",
            "properties": {
                "outputs": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "quantity_processed": {
                    "type": "integer"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}